bugfix: Outlet dies causing deadlock.
After the outlet made @max_reqs_per_conn requests
the #finish call would be made to the http client.
This caused an exception. Unless you are aborting thread
exceptions, this would cause your program to continue running
even though the outlet thread was dead. This ment that there
would be no process taking items out of the @request_queue. This
would cause the calls to #flush to block thus halting progress of the
entire program.