Make hashed queues thread safe.
Each queue now will expose an API to determine if it is full
and also an API to drain all messages from the queue.  This change does
not impact current behaviour and is a step forward towards creating fine
grained locks for lpxc such that we can expose a flush API for the client.