Feature: expose `flush` in API
Lpxc#flush is now thread safe. In order to make it thread safe, we had to
synchronize a mutex inside the method. This meant that we had to remove
the call to flush in Lpxc#puts outside of the mutex in which we initalized the
FlushableLockableSizedQueue. Thus avoiding deadlock.

In addition, this commit removed the external locking in Lpxc#delay_flush
since the flush method is now responsible for such activity.